requirements.txt
pillow>=10.0.0
PyYAML>=6.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
